acs_ex v0.3.8 ACS

Request router for CPE->ACS

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(arg)

Callback invoked to start the supervisor and during hot code upgrades.

Developers typically invoke Supervisor.init/2 at the end of their init callback to return the proper supervision flags.

Callback implementation for Supervisor.init/1.

start_link(session_handler, port, ip, ip6, opts \\ [])