acs_ex v0.3.8 ACS.CWMP.Parser

This module implements a simple wrapper around the cwmp_ex CWMP library parser. The request body is assumed to be a SOAP envelope containing one or more CWMP requests/responses.

Functions

init(opts)

Just an init that must be here.

parse(conn, type, subtype, headers, opts)

parse(Plug.Conn.t, String.t, String.t, List.t, Keyword.t) :: Plug.Conn.t

Parses text/xml requests as CWMP requests/responses.

The resulting CWMP requests and responses are returned so they can be set as the body_params field of the connection.