ACS.Handlers.ACS (acs_ex v0.3.18)

Handles CWMP requests coming in over the CWMP southbound interface.

Requests are parsed as SOAP CWMP envelopes and then used for the basis of the current provisioning tree. There are several possible request types as defined in the TR-069 specification. These are all just handled by the custom logic module.

Some are part of basic ACS functionality and are handled here.

This is the main ACS server loop corresponding to the old acs3.pm

Functions

call(conn, opts)

Callback implementation for Plug.call/2.

dispatch(conn, params)

handle_errors(conn, assigns)

When our cwmp parser thows, we end up here. So when some CPE sends bogus xml, or half a record or whatever else they might do, this is where we capture that parser error from CWMP.Parser.

init(opts)

Callback implementation for Plug.init/1.