ACS.Session.Script.Vendor (acs_ex v0.3.19)

This is the main vendor specific logic module, and the one you want to write if you have stuff that needs to happen in your specific environment.

In the real world you would write a new module that starts acs_ex using the child_spec, giving it the name of your handler module.

That would make acs_ex use your module when initiating the script session.

An example of this is available right next to this module.

Functions

session_filter(device_id, inform)

Callback implementation for ACS.SessionScript.session_filter/2.

session_start(session, did, inform)

Callback implementation for ACS.SessionScript.session_start/3.