acs_ex v0.3.8 ACS.Session.Supervisor

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

end_session(session_id)

init(session_module)

Callback invoked to start the supervisor and during hot code upgrades.

Developers typically invoke Supervisor.init/2 at the end of their init callback to return the proper supervision flags.

Callback implementation for Supervisor.init/1.

start_link(session_module \\ nil)

start_session(session_id, device_id, message)

start_session(session_id, device_id, message, fun)